feat: Add axe and shovel images to tools section
This commit adds the axe and shovel images to the tools section of the game.
The pickaxe section is temporarily commented out.
A new class "mikum" is added to define the height and cursor style for the axe and shovel sections.
The size of the dirt block is adjusted to "contain".